Extracted exportScene logic to utils/exportScene.js for better modularity
